New rust hello world template
