+++ al2/rust/hello/cookiecutter.json
@@ -1,8 +1,10 @@
 {
-    "project_name": "My Project",
-    "project_slug": "{{ cookiecutter.project_name.lower().replace(' ', '-') }}",
-    "architecture": ["x86_64", "arm64"],
-    "_copy_without_render": [
-      ".gitignore"
-    ]
-  }
+  "project_name": "My Project",
+  "project_slug": "{{ cookiecutter.project_name.lower().replace(' ', '-') }}",
+  "architectures": {
+    "value": []
+  },
+  "_copy_without_render": [
+    ".gitignore"
+  ]
+}
