+++ al2/rust/hello/{{ cookiecutter.project_slug }}/rust_app/Cargo.toml
@@ -6,13 +6,9 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-aws-config = "0.52.0"
-aws-sdk-dynamodb = "0.22.0"
-aws-smithy-client = { version = "0.52.0", features = ["test-util"] }
-lambda_http = "0.7.2"
-tokio = { version = "1.21.1", features = ["full"] }
 
-[dev-dependencies]
-aws-smithy-http = "0.52.0"
-aws-types = { version = "0.52.0", features = ["hardcoded-credentials"] }
-http = "0.2.8"
+lambda_runtime = "0.6.0"
+serde = "1.0.136"
+tokio = { version = "1", features = ["macros"] }
+tracing = { version = "0.1", features = ["log"] }
+tracing-subscriber = { version = "0.3", default-features = false, features = ["fmt"] }
