+++ manifest-v2.json
@@ -1050,6 +1050,14 @@
       "appTemplate": "hello-world",
       "packageType": "Zip",
       "useCaseName": "Hello World Example"
+    },
+    {
+      "directory": "al2/rust/hello-ddb",
+      "displayName": "DynamoDB Example",
+      "dependencyManager": "cargo",
+      "appTemplate": "hello-world",
+      "packageType": "Zip",
+      "useCaseName": "DynamoDB Example"
     }
   ],
   "amazon/nodejs14.x-base": [
